+++ react/components/PriceRange.js
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react'
 import PropTypes from 'prop-types'
 import { useRuntime } from 'vtex.render-runtime'
-import { injectIntl, intlShape } from 'react-intl'
+import { useIntl } from 'react-intl'
 import { Slider } from 'vtex.styleguide'
 import { formatCurrency } from 'vtex.format-currency'
 
@@ -13,8 +13,9 @@ import useSearchState from '../hooks/useSearchState'
 const DEBOUNCE_TIME = 500 // ms
 
 /** Price range slider component */
-const PriceRange = ({ title, facets, intl, priceRange }) => {
+const PriceRange = ({ title, facets, priceRange }) => {
   const { culture, setQuery } = useRuntime()
+  const intl = useIntl()
 
   const navigateTimeoutId = useRef()
 
@@ -95,10 +96,8 @@ PriceRange.propTypes = {
   title: PropTypes.string.isRequired,
   /** Available price ranges */
   facets: PropTypes.arrayOf(facetOptionShape).isRequired,
-  /** Intl instance */
-  intl: intlShape.isRequired,
-  /** Current price range filter query parameter*/
+  /** Current price range filter query parameter */
   priceRange: PropTypes.string,
 }
 
-export default injectIntl(PriceRange)
+export default PriceRange