+++ react/components/SelectedFilters.js
@@ -1,11 +1,11 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import { intlShape, injectIntl } from 'react-intl'
+import { useIntl } from 'react-intl'
+import { useCssHandles } from 'vtex.css-handles'
 
 import FacetItem from './FacetItem'
 import FilterOptionTemplate from './FilterOptionTemplate'
 import { facetOptionShape } from '../constants/propTypes'
-import { useCssHandles } from 'vtex.css-handles'
 
 const CSS_HANDLES = ['selectedFilterItem']
 
@@ -15,10 +15,10 @@ const CSS_HANDLES = ['selectedFilterItem']
 const SelectedFilters = ({
   map,
   filters = [],
-  intl,
   preventRouteChange = false,
   navigateToFacet,
 }) => {
+  const intl = useIntl()
   const handles = useCssHandles(CSS_HANDLES)
   if (!filters.length) {
     return null
@@ -56,11 +56,9 @@ SelectedFilters.propTypes = {
   map: PropTypes.string,
   /** Selected filters. */
   filters: PropTypes.arrayOf(facetOptionShape).isRequired,
-  /** Intl instance. */
-  intl: intlShape,
   /** Prevent route changes */
   preventRouteChange: PropTypes.bool,
   navigateToFacet: PropTypes.func,
 }
 
-export default injectIntl(SelectedFilters)
+export default SelectedFilters